# CTHz/openapi.yaml
openapi: 3.0.3
info:
  title: CTHz System 300 Device Management API (Internal)
  version: 0.1.0
  description: |
    Internal-only API specification for the System 300 Device Management (DM) service.

    Notes
    - This spec includes Spurtree-owned Device Management APIs and references CTHz-owned Sensor APIs (proxied by DM).
    - Do not expose Swagger/Redoc in production. Keep this file internal.
servers:
  - url: https://localhost:8443
    description: Local development (TLS)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          example: validation_error
        message:
          type: string
          example: Invalid request payload
        details:
          type: object
        requestId:
          type: string
          example: 9b2f9d62-4e1e-4a9b-95f1-3b9e0a9f3c1a
      required: [code, message]
    Ack:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, admin, monitor]
        is_active:
          type: boolean
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: User login authentication
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '200':
          description: JWT tokens
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
        '401': { description: Unauthorized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/auth/logout:
    post:
      tags: [auth]
      summary: User logout
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Ack' } } } }
        '401': { description: Unauthorized }
...
  /api/auth/setup:
    post:
      tags: [auth]
      summary: First-time setup (create Owner account)
      x-owner: Spurtree
      x-priority: High
      x-phase: MVP
      x-rbac: [owner]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        '201': { description: Owner created, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '409': { description: Already initialized, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/auth/me:
    get:
      tags: [auth]
      summary: Current user info
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      responses:
        '200': { description: Me, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { description: Unauthorized }
    put:
      tags: [auth]
      summary: Update user profile
      x-owner: Spurtree
      x-priority: Medium
      x-phase: MVP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { description: Unauthorized }
...